Extract AutoRebaseController from IssueTracker (CROW-1094)
Move the auto-rebase watcher (CROW-318) — applyAutoRebase, attemptRebase,
recordRebaseNoOp, deferRebase, the verdict/notify helpers, the eligibility /
backoff / escalation statics, and the AutoRebaseVerdict / AutoRebaseDeferReason /
AutoRebaseUpToDateHead / AutoRebaseDeferral types — into its own collaborator,
owning all its per-head bookkeeping. It reads the auto-merge branch-ownership
state (autoMergeInFlight / autoUpdateBranchAttempted / shouldUpdateBranchBeforeMerge)
and the shared codeBackend / prHasCrowAuthoredCommit / gitManager (widened to
internal, kept on the tracker) via an unowned back-reference. The dispatch Task
uses [weak self].

applyPRStatuses drives it; an extension IssueTracker re-exposes the
IssueTracker.<symbol> / tracker.<member> surface used by the tests and by
AutoReReviewController (shouldAttemptAutoRebase).

swift test --package-path Packages/CrowEngine: 789 tests pass.


Crow-Session: 18A4CD55-8B33-48E4-B5EB-E710314DF5E5
